+++ src/Admin/components/AdminHeader.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import { logo } from '../../../assets/index.js';
+import { logo } from '../../assets/index.js';
 
 const AdminHeader = ({ name, role, department, program, classname }) => {
     console.log(name, role, department, program);
