+++ src/Admin/index.js
@@ -1,0 +1,4 @@
+import Admin from './Admin.jsx';
+import Profile from './Profile/Profile.jsx';
+
+export { Admin, Profile };
