--- src/layouts/Admin/index.js
@@ -1,4 +1,0 @@
-import Admin from './Admin';
-import Profile from './Profile/Profile';
-
-export { Admin, Profile };
